add custom_field license into backoffice for all site which GobiertoData Module is enabled
