+++ db/migrate/20210413100034_add_to_dataset_custom_field_source_source_license.rb
@@ -1,0 +1,9 @@
+class AddToDatasetCustomFieldSourceSourceLicense < ActiveRecord::Migration[6.0]
+  def up
+    Site.find_each do |site|
+      GobiertoSeeds::GobiertoData::Recipe.run(site) if site.configuration.gobierto_data_enabled?
+    end
+  end
+  def down
+  end
+end
